Project Buildable
-project builds without errors or warnings
(aka all headers and src files are connected)

+++ STM32IDE_Examples/MainCode_Examples/Drivers/Bretts_Helper/Inc/OuterRing.h
@@ -5,7 +5,8 @@
  *      Author: Brett
  */
 
+#include <stdbool.h>
 
-int OuterRing_init(int *initalHitDis);
+int OuterRing_init(float *initalHitDis);
 
-int OuterRing_control(int distances[], int reason, bool initalDetachHit, int initalHitDis);
+int OuterRing_control(int distances[], int reason, bool initalDetachHit, float initalHitDis);
